Making grab work initially
